api/requirements.txt
flask
flask-cors
openai
redis
python-dotenv
gunicorn